buzz: cargo clippy
Do not fix clippy warnings for tests because it would give a false sense
of security. clippy rightly warns about comparing `f64`, but it does not
recognize that `Vec3`'s `PartialEq` implementation relies on comparing
`f64` too so fixing the warning wouldn't actually fix the problem.
Therefore, do not fix anything as a kind of reminder.