geo/kdtree: improve best_partitioning
Adopt a new strategy to find the best partitioning: find the median
X,Y,Z values for the centers. These values partition the space almost
equally by definition, but it's still possible to have the same median
value multiple times which can result in a non ideal partitioning. To
mitigate this issue, iterate over all the median values and find the one
that best partitions the input.