geo/kdtree: fix intersection by properly partitioning shapes
Shapes with a bounding box that encloses the partitioning point need to
be shared among the left and right sub-nodes because the intersections
are calculated by splitting the ray given (split_axis, split_value) and
checking the right and left sub-nodes for possible intersections using
the two new rays. However, if a shape that encloses the split_value only
appears either in a left or right sub-node then it's possible that the
shape intersects the other portion of the original ray.