buzz: make Object return an Hit object as Intersection
This allows Objects to have more flexibility in shading.

This required a change to the Shape interface in order to forward a
lifetime to work around the lack of Generic Associated Types (GAT).
Moreover, this also caused the use of a couple unsafe transmute calls
because Rust doesn't seem to infer the proper lifetimes.