buzz/csg: improve bbox
In case of union the bbox is simply the union of the bounding boxes. In
case of intersection the bbox is simply the intersection of the bounding
boxes. In case of difference... well, I'm not too sure. Right now, we do
a probably non optimal thing: return the bbox of the first element which
is always correct because a difference between objects can be as big as
the first object but no more.